# src/openapi.yaml
openapi: 3.0.0
info:
  title: Tankman API
  version: 1.0.0
  description: API for managing tanks and measurements
servers:
  - url: http://localhost:3001
    description: Local server
paths:
  /tanks:
    get:
      summary: Get all tanks
      responses:
        '200':
          description: A list of tanks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TankResponse'
    post:
      summary: Create a new tank
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TankRequest'
      responses:
        '200':
          description: The created tank
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TankResponse'
  /tanks/{id}:
    get:
      summary: Get a tank by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The tank with the specified ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TankResponse'
    put:
      summary: Update a tank by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TankRequest'
      responses:
        '200':
          description: The updated tank
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TankResponse'
    delete:
      summary: Delete a tank by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tank deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /measurements:
    get:
      summary: Get all measurements
      responses:
        '200':
          description: A list of measurements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MeasurementResponse'
    post:
      summary: Create a new measurement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeasurementRequest'
      responses:
        '200':
          description: The created measurement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementResponse'
  /measurements/{id}:
    get:
      summary: Get a measurement by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The measurement with the specified ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementResponse'
    put:
      summary: Update a measurement by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeasurementRequest'
      responses:
        '200':
          description: The updated measurement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementResponse'
    delete:
      summary: Delete a measurement by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Measurement deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /data/latestmeasurementbytankid/{id}:
    get:
      summary: Get the latest measurement by tank ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The latest measurement for the specified tank
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementResponse'
  /data/allmeasurementsbytankid/{id}:
    get:
      summary: Get all measurements by tank ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: All measurements for the specified tank
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MeasurementResponse'
  /data/livemeasurementbytankid/{id}:
    get:
      summary: Get a live measurement by tank ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A live measurement for the specified tank
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiveMeasurementResponse'
  /health/status:
    get:
      summary: Health check
      responses:
        '200':
          description: Health check status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatusResponse'
components:
  schemas:
    HealthStatusResponse:
      type: object
      properties:
        uptime:
          type: number
        message:
          type: string
        date:
          type: string
          format: date-time
    TankResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        diameter:
          type: number
        height:
          type: number
        statedCapacity:
          type: number
        capacityUnit:
          type: string
        timeStamp:
          type: string
          format: date-time
    MeasurementResponse:
      type: object
      properties:
        id:
          type: integer
        tankId:
          type: integer
        level:
          type: number
        availableCapacity:
          type: number
        availablePercentage:
          type: number
        timeStamp:
          type: string
          format: date-time
    LiveMeasurementResponse:
      type: object
      properties:
        level:
          type: number
        status:
          type: string
        timeStamp:
          type: string
          format: date-time
    TankRequest:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        diameter:
          type: number
        height:
          type: number
        statedCapacity:
          type: number
        capacityUnit:
          type: string
        timeStamp:
          type: string
          format: date-time
    MeasurementRequest:
      type: object
      properties:
        tankId:
          type: integer
        level:
          type: number
        availableCapacity:
          type: number
        availablePercentage:
          type: number
        timeStamp:
          type: string
          format: date-time